Set tls to true when on default and an option is set, remove description on tls/ssl
